+++ Templates/Tests/Context/AutoMockable.swift
@@ -39,6 +39,14 @@ protocol ExtendableProtocol: AutoMockable {
     func report(message: String)
 }
 
+protocol ReservedWordsProtocol: AutoMockable {
+    func `continue`(with message: String) -> String
+}
+
+protocol ThrowableProtocol: AutoMockable {
+    func doOrThrow() throws -> String
+}
+
 extension ExtendableProtocol {
     var canReport: Bool { return true }