+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Added support for file paths in `config` parameter
 - Improved config file validation and error reporting
+- Support methods with reserved keywords name in `AutoMockable` template
+- Support methods that throws in `AutoMockable` template
 
 ### Bug fixes